[FIX] website: work around webkit range bug around inline-blocks
<p>a[]b<inline-block>cde</inline-block></p> -> press End button
-> <p>ab<inline-block>cd[]e</inline-block></p> -> press Home button
-> <p>ab<inline-block>[]cde</inline-block></p>

This is a webkit bug. To work around it I edited the css of website in
edit mode so bootstrap .btn are not inline-block but preserve a similar
styling, as I wasn't able to identify another situation than bootstrap
.btn in which we also encounter the bug.

As a sidenote, navigation keys issues around inline-blocks in
contenteditable were reported to webkit... 7 years ago:
https://bugs.webkit.org/show_bug.cgi?id=105898 so there is little hope
for them to fix this.