[IMP] web: make One2many field readonly for multi_edit tree view
It does not make any sense if we are allowing to edit multiple rows having
one2many value changes in multi_edit list view.  It will lead to many confusion
to the user that changes multiple records by mistakenly.

So we are now restricting to edit the One2many field value having multi_edit
tree view. We are now changing the One2many field as read-only forcefully
when a user tries to edit the multiple row.

Task : 2150384