[MOV] point_of_sale: Move IoT Box code to separate repo
As the IoT Box use is not limited to the POS anymore, it doesn't make
sense to keep its code in the point_of_sale module.

Moving the code to a separate repository also allows us to disociate
the version of Odoo and the version of the IoT Box.

TaskID: 2333909