[FIX] website: solve long dropdown menu problems in navbar
Following [1]. The mentioned commit selectors were to be simplified and
actually did not solve the problem in mobile affixed menu. Also,
targeting "o_affix_enabled" to target the non-affix menu was wrong since
when there is no affix menu, no header has the "o_affix_enabled" class
as its names indicates it. So the mentioned commit fixed half of the
problem and only when the website used the affix menu feature... Let's
hope this new commit fixes everything.

[1]: https://github.com/odoo/odoo/commit/21959599d1c25321865afabe7f70191efdda154e

closes odoo/odoo#59597

X-original-commit: 1e3897d44a6454d1672d3e585e516381d77342c4
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>