[FIX] core: don't inherit the `init` flag on downstream module dependencies
When a module is marked as `to update`, it makes sense to also update all
its downstream dependencies. The same applies to the `demo` flag.

It's different for the `init` flag. A module is marked with the `init`
flag when using the `-i` command-line option, which doesn't include the
downstream dependencies.

During the upgrade of a database to a newer version, it happens that some
modules are split, like the module `uom` that has been extracted from
`product` in version 12. In order to correctly install these new modules,
they are marked as `init` (as if they were passed as `-i` option).
However, the downstream dependencies of these new modules (in our example,
`product` and all others installed modules that depend on it) should not
have this `init` flag.

When modules have this `init` flag, their data (from XML files) are loaded
in `init` mode (obviously), which ignores the `noupdate` flag of records,
forcing the update of records that shouldn't.

Note: The data of the modules in `to install` state are always loaded in
`init` mode.