[FIX] survey: ensure survey graph max height is based on most voted answer
Before this commit, the graph y scale in the survey manager screen was based on
the attendees count. But is some attendees do not answer the question, the bars
never reach the maximum height of the graph, and bars can stay really small.

Example: 10 attendees - 4 answers

10 |
   |
   |
   |       |
0  |_|__|__|_
     A  B  C

After this commit, the count of the most voted answer is used to set the height
scale of the graph. So the bars will always take full height space, relatively
to the count of the most voted answer.

Example: 10 attendees - 4 answers

4  |       |
   |       |
   | |  |  |
0  |_|__|__|_
     A  B  C

Task ID: 2351455