[FIX] website_sale: fix ribbon modifications not being previewed
Previously, editing a ribbon in the options panel had no preview inside
of the page until the change was validated. This commit fixes that issue
by making the real ribbon invisible, and making the ribbon dummy
visible. The preview still only affects the current ribbon being edited
until it is validated. Unvalidated changes are discarded when clicking
elsewhere in the page.

task-2339986

closes odoo/odoo#59782

X-original-commit: bdda4a06e2b0543f46cf51960af20505d574447b
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>