[IMP] web: allow js_class for FormViewDialog
Allowing js_class in FormViewDialog with the help of arch of the viewInfo and
adding it into the viewRegistry. So, user can simply add js_class in xml side
and build it's own structure.

For example- In planning module, gantt view has FormViewDialog so with the help
of js_class we can simply change the behaviour of footer button.

LINKS
Task- 2322036