[FIX] digest: send the digest tips 2 hours after the database creation
PURPOSE

We want to send the digest tips 2 hours after the database creation.
It can not be done directly on the digest itself because its field
is a date and not a datetime.

So we change the "Next Run Date" of the digest to now, and we run
the digest CRON 2 hours after its creation.

LINKS

Task 2351561
Internal odoo/internal/pull/1051

closes odoo/odoo#59474

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>