[FIX] crm: fix mail extension module documentation URL
The URL to the mail extension module documentation in the settings view is
currently wrong.
It only works temporarily because of a redirect we setup on our website.

This commit updates the URL to the proper location of the documentation.

When we later support Gmail, the view will have to be updated to contain two
documentation links: one to Outlook and one to Gmail.

Task 2356910

closes odoo/odoo#59481

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>