[FIX] hr_holidays: add explicit domain on action
In odoo/odoo@d2238de57630d58c4913ab44746476b83492c72f the domain of the
action hr_holidays.hr_leave_action_action_approve_department has been
removed, as the addon was reworked and no domain is needed any more.

However, during a migration from a previous version, the older domain is
not overwritten with an empty value.

With this commit, the domain is set to an empty one, which allows the
migration to correctly update the record.

closes odoo/odoo#59656

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>