[ADD] wowl: add loading indicator + loading overlay
Add component LoadingIndicator. LoadingIndicator displays a small rectangle at the bottom right with the number of current RPC requests. The rectangle is displayed if there is at least one current RPC request.
Add the ui service that allows to block and unblock the ui.
Add test for LoadingIndicator and uiService

Task-2358626