[FIX] core: set _rec_name field when creating from many2one field
Consider a many2one field such that the comodel's `_rec_name` is not
`'name'`.  In a form view with this field, type some text in the field
and click "Create and Edit...".  A form appears to create a new record.
And here is a problem: the comodel's `_rec_name` in the form is not
initialized with the text from the many2one widget.

The reason is the following: the form does not know the field's
comodel's `_rec_name`.  It actually opens the form with the text as
`default_name` in the context.

We propose the following fix.  As a model's `_rec_name` is the base
implementation of the field `display_name`, the client should open the
form with the text as `default_display_name` in the context, so that
part is now truly generic.  We also add a special case for this in
method `default_get`, such that it uses the context value as a default
value for the model's `_rec_name`.