[IMP] acc: Improve user errors in general
User errors in general could use some improvements to ease the flow.
This PR aims to do what is possible to improve them.

This is done mainly by changing their type from user error to redirect
error and adding a redirection to the related model, making it easier
to find where the change is needed.

This also hide the Cut-Off button on account move lines when the move is
 in draft, since it doesn't work for draft move lines anyway.

Task id #2345376