[IMP] wowl: Improve assets management
Before this commit, the method loading the assets files and folders from
the manifest was quite straightforward and allowed basic assets
declaration.

This commit introduces the work-in-progress function that will load the
assets in the future webclient, which allows singleton file declaration,
explicit external addon namespace, sub-bundle calls and the handling
of JS files.