[IMP] project: Add styles and override ListView class
In List view of tasks, a new button appears on each row. This button
allows to user to easily mark as done the task if it is not the case
yet. If the task is done then the button changes the kanban_state from
'done' to 'normal', this case allows the user to return back.

task-2381914