[FIX] stock: reserve return move
Usecase to reproduce:
- Send 2 LOT A and 2 LOT B.
- Return the picking with a quantity of 4.

The return picking is in ready state without the correct lot and
quantities on move lines. It happens because customer location is
not internal and bypass reservation.

This commit do not skip reservation in case of return. Origins moves
are already linked to the return to get the same lot and quantities than
ordered.