[IMP] project*: improve sub tasks creation and workflow
We want to ease user experience regarding sub task creation and
management. We thus enforce the values of subtask stage, project and
parent task. we want to remove the ability of a user to choose a
third-party subtask project different than the actual one. Also, we want
to remove the ability to make one sub task recurrent. In general, We
want to improve and strengthen the relation parent-childs for Tasks. A
sub task is dependant of its parent all along the workflow.

This commit address those needs and, especially, removes the subtask
project configuration discussed in :
- odoo/odoo@66a0e5a
- odoo/odoo@b709e22

1) With this commit we want to change the way to link a subtask with its
project	for UI reasons, we want ot be able to easily distinguish
subtasks that are linked to the project of their parent from subtasks
that are linked to another project. To do so, the idea is to make the
field project_id not required anymore. Only relevant tasks are linked
to their project, the other are linked to their parent which are linked
to a project. As a result, domains in the application are only based on
the fact that tasks have project_id. The other are simply hidden.

This is a threat to both security and reporting on several workflows.
Regarding security : Some Record Rules are based on the project
configuration. If project is not linked to a task, the rule is broken
and not relevant anymore. Regarding reporting : If we want to have all
tasks of a project, both parent and sub tasks, we have to use some
child_of queries which threaten the perfomances.
Also, with a smaller impact, we have to change some related fields in
the task model into computed fields, again a loss of performance.

To cover those losses, a new field has been created : the
effective_project_id which will be the field to use for security
reasons or consolidation of tasks with projects.

2) The stage of a task is the same as the one of its parent if project
   is not set.

3) The number of subtasks is shown on kanban and list views with a
   wigdet to easily render it and extend the UI.

4) The default sub-task assignee is the assignee of the parent task.

5) With this commit we want to remove the subtask global setting.
   Sub-tasks are enabled by default but can be disabled by project.

6) Task form is modified to integrate those changes.

PR Community: #62577
PR Enterprise: odoo/enterprise#15043
PR Upgrade: odoo/upgrade#1984

task-2388034
Part of task-2379043