[FIX] sale_timesheet: remove the warning messages in project.task
When the user changes the sale order line or project in a task,
we displayed a warning message about the impact of this changes on
the timesheets in the task.
This warning message was useful to inform the user that is a change
in the behaviour compared to the previous version of Odoo.
But each time, the warning message appears and it can be boring for
the user. That's why we remove the warning message.

task-2388500