[IMP] sale_timesheet: add task_id in api.constraint
Before this commit the _check_sale_line_in_project_map method used the
task_id field in the condition to check if the timesheet can be billed,
however, this field is not in the api.constraint decorator of this
method.

This commit adds this field in the api.constraint and directly filter
the timesheet before to enter in the for loop.

task-2388500