[IMP] project : View subtask as subrows in a task tree view
[Change aborted]
We want to list subtasks in the task tree view to be able to easily show subtsks with the same columns as its parent.

With this commit, sub tasks are fetched with their parent task in the tree view and can be collapsed/expanded by the user.
The user has still the ability to add, show or hide columns, to group tasks, the behaviour is the same for parent task and subtasks.

Task-2388034 : Part of task-2379043