[IMP] event_sale: adapt test and tour for event_booth modules
PURPOSE

When confirming a sale.order that includes a product that is either an
event registration or an event stand it shows the event edit registration
wizard which has no sense for the event stand part.
Fix the event_configurator_tour since the autocomplete showed the wrong
product at the first place.

SPECIFICATION

The test was based on the event_id field of the sale.order.line but should
be based on the event_ok field instead.
Add more text to the tour so the autocomplete shows the right product at
the first place

LINKS

Task ID : 2318655
PR : #62759