[IMP] sale_timesheet: remove the context key added
This context key has been added in the 14.0 for a fix in helpdesk to
create a sale order form a ticket.
But, now we remove the 'create sale order' button in the ticket to add
sale order line field in the ticket, this field allows to link the
timesheet in this sale order line selected in this field.

Thus, this commit removes the context key used 'ticket_timesheet_ids'
since this key will never be given again in this wizard.

task-2388500