[IMP] project*: improve project subtasks
This commit reverts the removal of parent_id field in task form to allow
user to still change the task
inheritence after task creation.

This commit adds optional columns in the one2many subtask field in the
task form.

This commit adds an on-hover behaviour to assign/unassign button to
improve the UX. It adds a tree
class and a field class to easily show/hide element in a tree view.

This commit shows sub task editable fields in the one2many tree view.
subtask>project_id and subtask>stage_id does not take the
default_project_id/default_stage_id form value as those are computed.

This commit adds the assignee of the parent task as default assignee of
subtask.

This commit aims to change the widget displaying the number of subtasks
in kanban and list views.

This commit aims to hide task with project_id not set in various views.

The subtasks now rely on the project of their parent and their
project_id is not set. Thus, by adding this condition in the domain of
those various views, the subtasks are hidden.

Various searches on project task are restricted with a clause where
project_id must be not null.

PR community: #62577
PR enterprise: odoo/enterprise#15043
PR upgrade: odoo/upgrade#1984

task-2388034