[FIX] website_forum: disable image resizing in editor toolbar
Before this commit, the inline style added on image when using the
resize options in the editor toolbar of a post, was removed due to the
strip_style of the content field.

Since we cannot remove this strip_style for security reasons, we can
only disable the resizing options.

ISSUE 34288