[ADD] sale_timesheet_edit: allow edition of so_line in timesheet
This module allows the user to edit the so_line of each timesheets in
a task in project.task form view.
When the user manually edits the so_line of a timesheet, we need to
store in is_so_line_edited that this field has been edited manually.
We need to keep track this information because we need to keep the
so_line selected by the user even if the configuration of the project
changes or if the sale_line_id of the task changes.

task-2409761