[IMP] mail: unfollow button in discuss
PURPOSE

It would be nice to have an unfollow button when in the inbox, starred and
history. This way, if we receive a notification for a record we no longer are
interested in, we could easily stop this from happening again.

SPECIFICATION

Adding an unfollow button with icon 'fa-bell-slash-o' next to the name of the
record, should only be visible if the user is part of the followers from the
thread. Clicking on the icon, the notification will be marked as read and the
user will be removed from the followers of the thread.

LINKS

PR https://github.com/odoo/odoo/pull/67759
Task- 2242170