[IMP] web, sms: clickable area for url widget is limitated and phone widget test case changed
as the clickable area issue is also found in url widget same as email widget
we solved the issue by wraping the anchor tag within a container div tag. here,
the overflow class will be on the container div which will do it's
job to prevent the long strings from breaking the UI, and the anchor
tag being its child, will not be the full width, thus limiting the
clickable area.

and in this commit some qunit test cases for phone widget is also changed as it is
extended from email widget.

task-id: 2345974