[FIX] pos_discount: applying global discount should be for each line
Current application of global discount involves creating new line
which has the 'discount' product and a unit price computed based
on the order amount. However, this method is flawed when the lines
involved have taxes. The discount line can't take into account the
taxes distribution especially when each line have different taxes.

In this commit, we propose a new way of applying discount by setting
the discount to each line. With this method, taxes are computed
correctly regardless of the tax configuration for each product in
each line.