[FIX] mail: allow to use the mail composer in mass mode with template
Purpose
=======
Now, standard users can not render arbitrary Jinja code. But they still
can load mail templates which contains Jinja and they will be rendered
if the mail composer is in *comment mode*.

But in *mass mode*, we can't load template
because many things change the body.
- Run HTML sanitizer on the body field before saving the composer
  (sanitize=False on the <mail.template> but True on the <mail.compose.message>)
- Add signature (13.0 and older versions)
- Web editor change the body HTML (move CSS properties, add spaces, etc)

So this behavior break some flow where the mail composer can be used in
mass mode so send email based on a template
(e.g.: message_post_with_template). Ideally we should be able to render
the body if it didn't change in comparison with the template.

So after this commit, non-"Mail Template Editor" still won't be able
to render template in the mail composer in mass mode from the frontend
(because the web editor will change the body) but the methods which call
the mail composer with a mail template in mass mode will work (because
the body won't be changed by the web editor).

Task 2187263