[IMP]crm: no need for revenues to get rainbows
when there is no excpected revenues is set on opprtunity.
then the rainbowman is display for the "first won". which
is correct behavior.

In this module if there is no excpected revenues is set on
opportunity. then no rainbowman is display for the
"firstwon". While it does make sense to ask for a
revenue so that we can measure if this is a record or not,
it is not necessary to congratulate the user for his first
win. so we remove the condition excepected_revenue for this
spesific case.

with this commit, we remove the specific condition which is
excpected_revenue for this specific case. so now if there is
no excpected revenues is set for opprtunity. the rainbowman
will be display for the "firstwon".

TaskID - 2410443