[ADD] mail_template_security: restrict the access to the templates
Purpose
=======
Prevent non "Template Editor" from writing on the mail / SMS templates
and do not allow to render non-trusted Jinja content.

Specifications
==============
Create a new group "Mail Template Editor".

The creation / writing of the Mail / SMS templates are now restricted
to this group.

Prevent users who are not in this group to render a Jinja content which
do not come from a trusted source (Mail / SMS template).

Task 2187263