[FIX] Digest: Encoding digest images in base 64
Problem:

The digest will contains images that are hosted on the odoo
instance of the client. If the instance is deleted, the mails
will contain dead links and the images will no longer be able
to be loaded.

Solution:

The images of the digests will now be encoded in the mail.
With this approach, the images can be diplayed without having
to load them from a remote server. Here, we used this approach
to avoid maintaining a service that serve images.