[IMP] web: add a basic progressbar widget
=============================================================

Disclaimer : The purpose of this PR may be ambiguous.
The first objective was to make the progressbar editable with the click.
Given the current implementation of the progressbar widget, an
inheritence wasn't possible to achieve it. Technically, the init
function of the progressbar cannot be inherited without re-writing it
completly. Thus, the first development lead to a new widget with the
purpose to have an click-editable progressbar.

This commit may change the objective of the PR and could lead to a
refactoring of the progressbar widget. Yet, the developer (tle) didn't
take the responsability to directly trash the progressbar, as it was not
meant in the first specification of the task. It may be
interesting to discuss it between the js framework guru and the PO.

==============================================================

This commit aims to add a new progress bar called basic_progressbar
which, basically, allows user to see integer or float represented on a
progressbar.

This wigdet is configurable (see node options in the widget
specifications) and with default options, it :
- Allows user to edit value through an input.
- Shows the value as a percentage.
- The max value of the bar is 100%.
- The number is formatted depending on its type (integer or float)
Developer may want to tune it and use it :
- with the default maximum value spelled out ("/ 100") and not
  represented as a percentage : 'percentage' (boolean);
- with a different maximum value : max_value_num (a number),
  'max_value_field' (a field present in the view);
- with an editable progressbar (with the mouse or the arrows) :
  'drag_and_drop' (boolean);
- with a human number format : 'human_number' (boolean);
- with a different than default format type : 'format_type' (string e.g.
  "float" or "integer");
- to show the field as the maximum value and not as the progress value :
  'edit_max_value' (boolean)
- with this last option ('edit_max_value') and use a other field as
  progress value : 'progress_value_field' (a field present in the view)

Technical comment :
- The _doAction function is called when the input change or when the
  progressbar is updated with the click or the keyboard arrows.
- The _allowEdit functions allows other developpers to tune the
  conditions which allows one to edit the widget in some views/context.
  As it is used with a logical OR, condition can only add cases and not
  restrict the default one.

------------------------------------------------------------------

Furthermore, this commit adds a new quick_edit_progressbar which
inherits from the basic_progressbar and can be used in a readonly
context such as kanban views or list views.

This quick_edit_progressbar can be edited once the user click on it and
returns to readonly mode once the user either blur the input or click
outside the quick_edit_progressbar widget.

Technical comment :
- The value is updated (sent to server by rpc) only when the field is
blurred, this behaviour could be tuned for further needs by overriding
_valueChanged. The behaviour is a design choice.
- The window click listener is only added if the widget is edited once
and is not re-added if the widget is edited-blurred-edited. Furthermore
this listener is only added when using the drag and drop option with
this widget.

------------------------------------------------------------------

PR : #64239
task-2412223