[IMP] delivery: formatted the field positioning under 'Sales and Purchase'
Currently when we install 'delivery' module, in the partner form a new
field 'Delivery Method' is added under 'Sales Person' field under the tab
of 'Sales and Purchases' and results in addition of extra fields between
Sales person and Sales Team , which is a bit awkward in the view.

In this improvement we intent to manage those fields to group together
by rearanging field's positions which is coming in between them.

Task Id : 2445997