[FIX] web: Auto close dropdown is break
Since odoo/odoo@e4f87710e16357f24f388159a7b1e82eccda598a, we added a way
to prevent bs dropdown and owl dropdown to be open in the same time.

But when the web editor is existing on the screen, the css selector used
to catch a existing modal opened match with an existing css selector
used inside the web-editor (it's conflict with web-editor modal
note-picture-dialog, note-link-dialog, note-help-dialog). To avoid this
sad behavior, this commit use now a more restrictive css selector to
catch only an existing opened dialog as a first child of the body.

Steps to reproduce:
1. Go to Project App
2. Open and edit a task
3. Click on Customer => search more
5. Click on "Filters"
6. Click on "Group by"
=> The "Filters" and "Group by" dropdown are open at same time => bug