[FIX] crm: tour vs activity popup
During this step of the tour, we want the user to drag
and drop the card. But one should still be able to schedule
an activity. The tour should never prevent you from doing something.

In this commit the possition of the bubble is changed from
bottom to the right of the card, thus it will not prevent
users from scheduling activities.

task-id: 2449223