[IMP] web: keep notebook page when enabling the studio
currently, When enabling studio on a form view, it reset the notebook page
and activate the first page instead of the currently activated page.

this commit fixes the issue by adding autofocus on the activated node
and therefore enabling studio on a form view, it will display the currently
activated page instead of the first page.

TaskID-2416755