[IMP] base : display partner_ids in wizard
if we are merging multiple contacts with the same name,
the destination contact field might be confusing as it just
says the name and does not include the id,
or other identifier to make it clear the exact contact you are
setting as the destination.

so in this commit we have added the partner ids along with its
name in destination contact field which will avoid the confusion
while merging contacts.

task-id: 2323060