[FIX] properly record history on snippet option select
/!\ This reveals bugs in the history generated by the historyStep
triggered in a setTimeout in the mutation observer's callback. History
therefore appears more broken after this commit than before it. Removing
the aforementioned historyStep fixes all that but that would break the
things that step was there to fix. I suggest we remove that historyStep
anyway and find a better solution because the code this commit changes
is a time bomb if left unchanged.