[REF] stock: test_stock_flow and test_warehouse + update quantity wizard
Note quantity wizard is for when Updating Quantity when no tracking,
multi-location, packages, or consignment. Also maybe fixed a mistake
where it was product packagings rather than packages...

TODO: Maybe instead of adding action_apply to wizard, it can be added to
write. Needs to be inspected further if this is possible.