[FIX] website_event_track: adapt website menu tests
PURPOSE

With the addition of website_event_booth we add a new website.event.menu
which is set by default along with the website_menu field. It messes the
tests when comparing the actual menu size.

SPECIFICATIONS

We specifically set the website_booth field of the event to False when
present in the model to fix the tests.

LINKS

Task ID : 2318655
PR : #62759