[IMP] web_editor, *: add a replace media button as an opt in the panel
*: website

On top of the new overlay button to replace an image, we now have a
snippet option button. This is a duplicate of the editor one which is
meant to be removed in the future... hopefully.

Note 1: this is not enabled on icons for now even if the code has been
        made for them as well. There is a bug which is likely to be
        solved with the new editor so best not waste time on it now.

Note 2: there are some glitches when switching between different media
        types (image/icons/video). Again, this is likely to be solvable
        by the editor team so best not waste time there as well.