[FIX] (test_)mail(_full,_enterprise): update status failure notification instantly
Before this commit:

Currently, failure notifications are sent to the author of the message that the
failure is concerning.

If another user than the author is resolving the failure, he is not receiving
any notification without refresh the page, so it appears as if nothing happened
even though the failure was indeed resolved

After this commit:

Failure notifications will be also sent to the partners of the message instantly
that the failure is concerning and also update the status of failure
notification instantly, no refresh needed to show the updated status.

Task-2178257