[IMP] website: refactored action_bw_check_assets
The function now behaves in a more generic way, allowing for table
schemas that dont have website_id or type columns

This means that non qweb entries of ir_ui_view will be considered, but
it does not seem to cause problems.

Also fixed the regex, which was not entirely proper and missed some
valid data.

Added a nested pretty print function for debug purposes.