[FIX] web: support for SQL dates
Before this commit, when trying to parse a date time that came directly
from the SQL ("yyyy-MM-dd HH:mm:ss"), the parseDateTime function
answered with an Error.

Now, the SQL format is accepted and parsed. Note that this format is
almost identical than the ISO one, the only difference is the lack of
'T' on the format ("yyyy-MM-ddTHH:mm:ss")