[FIX] mrp_account: duplicated analytic account entries
Previously, we always call _cost_generate to post analytic entries at
the end of button_mark_done, no matter the MO is done or not. In some
cases, for example when backoroder is needed, the state of the MO won't
be "done" after the first time we call button_mark_done. But the
analytic account entries will be post, and it will be post again after
we confirm we need backorder or not.
To fix, don't post analytic account entries for MO that is not "done".

Task-2469742
PR #68708