[IMP] project: add rightpanel in project update view
This commit adds a right panel in the project update view.

This is done by extending both KanbanView and KanbanController.
We extend KanbanView to add a RightPanel config, with the RightPanel
component to use.
In the KanbanController, we use the component Adapter to launch the Owl
RightPanel Component from a Legacy Controller.
This solution is based on SearchPanel and ControlPanel implementation.

The Project RightPanel Component gets information about the active
project in the context through a rpc call and render the resulting data
in a owl template. This template doen't have the ambition to be generic
but only specific to this view.

Serverside, the data are collected in a dict through few methods in
order to be easily extended.

PR: #68899
task-2393768