[REM] hw_drivers: listen and settimeout aren't needed
The timeout only applies to the listening socket that's used to create
new sockets through its `accept` call in SocketManager's run method.
Since that's all the `run` method does, timing out of the `accept` call
and looping around right back into that `accept` call again isn't
needed.

Likewise, the effect of restricting the listen backlog to 1 is that if
multiple devices connect at the same time, some of them will be rejected
instead of being accepted in turn. There's no need to do this and the
OS default will suffice.